requirements.txt
# Core utilities
python-dotenv==1.0.1
certifi==2024.2.2

# Data science essentials
numpy
pandas
scikit-learn==1.4.1.post1

# Visualization
matplotlib==3.8.3
seaborn==0.13.2

# MongoDB
pymongo
imblearn

pyaml






#-e .
